@@ -1,58 +1,15 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Migrations\NexplorerDbMigrations\**" />
-    <EmbeddedResource Remove="Migrations\NexplorerDbMigrations\**" />
-    <None Remove="Migrations\NexplorerDbMigrations\**" />
+    <PackageReference Include="MongoDB.Driver" Version="2.9.2" />
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Remove="Migrations\20180219191955_AddAggregateDataToAddress.cs" />
-    <Compile Remove="Migrations\20180219194741_AddAggregateDataToAddress.cs" />
-    <Compile Remove="Migrations\20180219200734_AddAddressAggregateData.cs" />
-    <Compile Remove="Migrations\20180219204516_AddAddressAggregateData.cs" />
-    <Compile Remove="Migrations\20180220195633_UpdateAddressRelationships.cs" />
-    <Compile Remove="Migrations\20180415100009_AddedUniqueConstraintsToBlockTxAddr.cs" />
-    <Compile Remove="Migrations\20180415100551_AddUniqueIndexAddressHash.cs" />
-    <Compile Remove="Migrations\20180506183954_RemoveFirstBlockRequired.cs" />
-    <Compile Remove="Migrations\20180506191556_RemoveFirstBlockRequired.cs" />
-    <Compile Remove="Migrations\20180506192317_RemoveFirstBlockRequired.cs" />
-    <Compile Remove="Migrations\20180506193140_RemoveFirstBlockRequired.cs" />
-    <Compile Remove="Migrations\20180506193913_RemoveFirstBlockRequired.cs" />
-    <Compile Remove="Migrations\20180506194155_RemoveFirstBlockRequired.cs" />
-    <Compile Remove="Migrations\20181101142017_RemovedOnDelete.cs" />
-    <Compile Remove="Migrations\20181101142344_RemovedRelationships.cs" />
-    <Compile Remove="Migrations\20181101142708_RemovedContraints.cs" />
-    <Compile Remove="Migrations\20181101142922_RemovedContraints.cs" />
-    <Compile Remove="Migrations\20181101143056_RemovedContraints.cs" />
-    <Compile Remove="Migrations\20181101184855_InitialCreateNoIndexes.cs" />
-    <Compile Remove="Migrations\20181101185559_InitialCreateNoIndexes.cs" />
-    <Compile Remove="Migrations\20181124144225_AddBlockRewardTable.cs" />
-    <Compile Remove="Migrations\20181124150230_AddBlockRewardColumn.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Include="Migrations\NexplorerDbMigrations\20181026123554_InitialCreate.cs" />
-    <Compile Include="Migrations\NexplorerDbMigrations\20181026123554_InitialCreate.Designer.cs" />
-    <Compile Include="Migrations\NexplorerDbMigrations\NexplorerDbModelSnapshot.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="8.0.0" />
-    <PackageReference Include="Dapper" Version="1.50.5" />
-    <PackageReference Include="StackExchange.Redis.StrongName" Version="1.2.6" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Nexplorer.Config\Nexplorer.Config.csproj" />
-    <ProjectReference Include="..\Nexplorer.Core\Nexplorer.Core.csproj" />
-    <ProjectReference Include="..\Nexplorer.Domain\Nexplorer.Domain.csproj" />
-    <ProjectReference Include="..\Nexplorer.Infrastructure\Nexplorer.Infrastructure.csproj" />
-    <ProjectReference Include="..\Nexplorer.NexusClient\Nexplorer.NexusClient.csproj" />
+    <ProjectReference Include="..\Nexplorer.Nexus\Nexplorer.Nexus.csproj" />
   </ItemGroup>
 
 </Project>
