@@ -1,77 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>bin\Debug\netcoreapp2.1\Nexplorer.Web.xml</DocumentationFile>
-    <NoWarn>1701;1702;1705; 1591;</NoWarn>
-  </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Resources\Scripts\External\**" />
-    <Compile Remove="Scripts\**" />
-    <Compile Remove="wwwroot\images\**" />
-    <Content Remove="Resources\Scripts\External\**" />
-    <Content Remove="Scripts\**" />
-    <Content Remove="wwwroot\images\**" />
-    <EmbeddedResource Remove="Resources\Scripts\External\**" />
-    <EmbeddedResource Remove="Scripts\**" />
-    <EmbeddedResource Remove="wwwroot\images\**" />
-    <None Remove="Resources\Scripts\External\**" />
-    <None Remove="Scripts\**" />
-    <None Remove="wwwroot\images\**" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Content Remove="nlog.config" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="Resources\Markup\pager-template.html" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Dapper" Version="1.50.5" />
-    <PackageReference Include="MailKit" Version="2.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.2.1" />
-    <PackageReference Include="Microsoft.AspNetCore.SignalR" Version="1.1.0" />
-    <PackageReference Include="NLog.Web.AspNetCore" Version="4.8.0" />
-    <PackageReference Include="protobuf-net" Version="2.4.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
-    <PackageReference Include="System.Reactive.Linq" Version="4.1.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Nexplorer.Config\Nexplorer.Config.csproj" />
-    <ProjectReference Include="..\Nexplorer.Core\Nexplorer.Core.csproj" />
-    <ProjectReference Include="..\Nexplorer.Data\Nexplorer.Data.csproj" />
-    <ProjectReference Include="..\Nexplorer.Domain\Nexplorer.Domain.csproj" />
-    <ProjectReference Include="..\Nexplorer.Infrastructure\Nexplorer.Infrastructure.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="App_Data\" />
-    <Folder Include="Resources\Images\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="nlog.config">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Update="Views\Admin\Index.cshtml">
-      <Pack>$(IncludeRazorContentInPack)</Pack>
-    </Content>
-    <Content Update="Views\Mining\Index.cshtml">
-      <Pack>$(IncludeRazorContentInPack)</Pack>
-    </Content>
-  </ItemGroup>
 </Project>
