@@ -1,104 +1,37 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-using Nexplorer.Config;
-using Nexplorer.Core;
-using Nexplorer.Data.Query;
-using Nexplorer.Domain.Dtos;
-using Nexplorer.Domain.Enums;
-using Nexplorer.Web.Cookies;
-using Nexplorer.Web.Extensions;
+using Microsoft.Extensions.Logging;
 using Nexplorer.Web.Models;
 
 namespace Nexplorer.Web.Controllers
 {
-    public class HomeController : WebControllerBase
+    public class HomeController : Controller
     {
-        private readonly BlockQuery _blockQuery;
-        private readonly TransactionQuery _txQuery;
-        private readonly RedisCommand _redisCommand;
-        private readonly StatQuery _statQuery;
+        private readonly ILogger<HomeController> _logger;
 
-        public HomeController(BlockQuery blockQuery, TransactionQuery txQuery, RedisCommand redisCommand, StatQuery statQuery)
+        public HomeController(ILogger<HomeController> logger)
         {
-            _blockQuery = blockQuery;
-            _txQuery = txQuery;
-            _redisCommand = redisCommand;
-            _statQuery = statQuery;
+            _logger = logger;
         }
 
-        public async Task<IActionResult> Index()
-        {
-            var channelStats = await _statQuery.GetChannelStatsAsync();
-
-            return View(new HomeViewModel
-            { 
-                LastBlock = await _blockQuery.GetLastBlockAsync(),
-                LastPosDifficulty = channelStats?.FirstOrDefault(x => x.Channel == BlockChannels.PoS.ToString())?.Difficulty ?? 0,
-                LastPrimeDifficulty = channelStats?.FirstOrDefault(x => x.Channel == BlockChannels.Prime.ToString())?.Difficulty ?? 0,
-                LastHashDifficulty = channelStats?.FirstOrDefault(x => x.Channel == BlockChannels.Hash.ToString())?.Difficulty ?? 0
-            });
-        }
-
-        public IActionResult Error()
-        {
-            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-        }
-
-        public IActionResult Cookie()
+        public IActionResult Index()
         {
             return View();
         }
 
-        [HttpPost]
-        public IActionResult DismissCookieWarning()
+        public IActionResult Privacy()
         {
-            var cookie = Request.GetCookie<UserSettingsCookieData>();
-
-            cookie.DismissedCookiePolicy = true;
-
-            Response.SetCookie(cookie);
-
-            return Ok();
+            return View();
         }
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult Search(string searchTerm)
+        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
+        public IActionResult Error()
         {
-            if (string.IsNullOrWhiteSpace(searchTerm))
-                return RedirectToAction("index");
-
-            searchTerm = searchTerm.Replace(",", "").Trim();
-
-            var isNumber = int.TryParse(searchTerm, out var height);
-
-            if (isNumber)
-                return RedirectToRoute("blocks", new {blockId = height});
-
-            switch (searchTerm.Length)
-            {
-                case 51:
-                    return RedirectToRoute("addresses", new { addressHash = searchTerm });
-                case 128:
-                    return RedirectToRoute("transactions", new { txHash = searchTerm });
-                case 256:
-                    return RedirectToRoute("blocks", new { blockId = searchTerm });
-                default:
-                    if (searchTerm.Length >= 20)
-                        return RedirectToRoute("blocks", new { blockId = searchTerm });
-
-                    return RedirectToAction("index");
-            }
-        }
-
-        [HttpGet]
-        public async Task<JsonResult> GetLatestTimeStampUtc()
-        {
-            return Json(await _redisCommand.GetAsync<DateTime>(Settings.Redis.TimestampUtcLatest));;
+            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
         }
     }
 }
