@@ -1,185 +1,8 @@
-﻿@using Nexplorer.Domain.Enums
-@model HomeViewModel
-
-@{
-    ViewBag.Title = $"#{Model.LastBlock.Height}";
-    Layout = "_Layout";
+﻿@{
+    ViewData["Title"] = "Home Page";
 }
 
-@section Head {
-    <link rel="stylesheet" href="@ViewBag.ControllerCss" />
-}
-
-@section Header {
-    <div id="hero" class="jumbotron jumbotron-fluid"></div>
-    <div id="heroText" class="jumbotron jumbotron-fluid">
-        <div class="container-fluid max-width">
-            <div id="heroHeader">
-                <h1>NEXPLORER</h1>
-                <p class="lead">
-                    A cryptocurrency block explorer for 
-                    <a href="https://nexusearth.com/" target="_blank">
-                        Nexus
-                        <span class="d-none d-sm-inline-block">
-                            (NXS)<img class="nxs-icon" src="img/nxs-icon.png"/>
-                        </span>
-                    </a>
-                    <br />
-                    <em>nex·usˈneksəs' : a connection or series of connections linking two or more things.</em>
-                </p>
-            </div>
-
-            <form id="mainSearch" asp-controller="home" asp-action="search" method="post">
-                <div class="form-group">
-                    <div id="mainSearchBar">
-                        <span class="fa fa-search"></span>
-                        <input class="form-control" name="searchTerm" type="text" />
-                        <button type="submit" href="#" class="btn btn-flat">
-                            GO
-                        </button>
-                    </div>
-                    <span class="instructions">Search for blocks, transactions or addresses</span>
-                </div>
-            </form>
-        </div>
-    </div>
-}
-
-<div class="panel full-length">
-    <div class="row d-sm-none">
-        <div class="col-9">
-            <h2>LATEST PRICE</h2>
-        </div>
-        <div class="col-3 text-right">
-            <span class="fa fa-btc"></span>
-        </div>
-    </div>
-    <div class="row">
-        <div class="col-sm-2 d-none d-sm-block">
-            <h2>LATEST PRICE</h2>
-        </div>
-        <div class="col-sm-3">
-            <h3>Last</h3>
-            <p class="val" v-cloak>
-                {{ lastPrice }}
-                <span class="fa fa-btc inline-fa"></span>
-            </p>
-        </div>
-        <div class="col-sm-3">
-            <h3>Spread</h3>
-            <p class="val" v-cloak>
-                {{ spread }} sats
-            </p>
-        </div>
-        <div class="col-sm-3">
-            <h3>Volume</h3>
-            <p class="val" v-cloak>
-                {{ volume }}
-                <span class="fa fa-btc inline-fa"></span>
-            </p>
-        </div>
-        <div class="col-sm-1 text-right d-none d-sm-block">
-            <span class="fa fa-btc"></span>
-        </div>
-    </div>
+<div class="text-center">
+    <h1 class="display-4">Welcome</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
 </div>
-
-@{ await Html.RenderPartialAsync("_AdvertBanner"); }
-
-<div class="row">
-    <div class="col-lg-6">
-        <div class="panel">
-            <div class="row">
-                <div class="col-9">
-                    <h2>LATEST BLOCKS</h2>
-                </div>
-                <div class="col-3 text-right">
-                    <span class="fa fa-cubes"></span>
-                </div>
-            </div>
-            
-            <div id="blockTable" style="margin-top: 15px;" v-cloak>
-                <table class="table table-hover" style="border: 0;">
-                    <thead>
-                    <tr>
-                        <th>Height</th>
-                        <th>Channel</th>
-                        <th class="d-none d-sm-table-cell">Size</th>
-                        <th><span class="d-none d-sm-block">Txs</span><span class="d-sm-none">TXs</span></th>
-                        <th style="width: 100px">Age</th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    <tr v-for="block in blocks" v-on:click="selectRow('/blocks/' + block.height)" style="cursor: pointer;">
-                        <td><a v-bind:href="'/blocks/' + block.height">{{ block.height.toLocaleString() }}</a></td>
-                        <td>{{ $layoutHub.parseBlockChannel(block.channel) }}</td>
-                        <td class="d-none d-sm-table-cell">{{ $layoutHub.parseBytes(block.size) }}</td>
-                        <td>{{ block.transactionCount }}</td>
-                        <td style="width: 100px"><date-age :from-date="block.timestamp" :precise="true"></date-age></td>
-                    </tr>
-                    </tbody>
-                </table>
-            </div>
-
-            <div class="text-right">
-                <a asp-action="index" asp-controller="blocks">
-                    <strong>More info</strong>
-                </a>
-            </div>
-        </div>
-    </div>
-
-    <div class="col-lg-6">
-        <div class="panel">
-            <div class="row">
-                <div class="col-9">
-                    <h2>LATEST TRANSACTIONS</h2>
-                </div>
-                <div class="col-3 text-right">
-                    <span class="fa fa-exchange"></span>
-                </div>
-            </div>
-            
-            <div id="transactionTable" style="margin-top: 15px;" v-cloak>
-                <table class="table table-hover">
-                    <thead>
-                    <tr>
-                        <th>Type</th>
-                        <th class="text-right">Amount</th>
-                        <th class="d-none d-sm-table-cell text-center">In</th>
-                        <th class="d-none d-sm-table-cell text-center">Out</th>
-                        <th>Age</th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    <tr v-for="tx in txs" v-on:click="selectRow('/transactions/' + tx.transactionHash)" style="cursor: pointer;">
-                        <td><a v-bind:href="'/transactions/' + tx.transactionHash">{{ $layoutHub.parseTxType(tx.transactionType) }}</a></td>
-                        <td class="text-right">{{ tx.amount.toLocaleString() }} <span style="font-size: 10px;">NXS</span></td>
-                        <td class="d-none d-sm-table-cell text-center">{{ tx.transactionInputCount.toLocaleString() }}</td>
-                        <td class="d-none d-sm-table-cell text-center">{{ tx.transactionOutputCount.toLocaleString() }}</td>
-                        <td><date-age :from-date="tx.timestamp" :precise="true"></date-age></td>
-                    </tr>
-                    </tbody>
-                </table>
-            </div>
-            
-            <div class="text-right">
-                <a asp-action="index" asp-controller="transactions">
-                    <strong>More info</strong>
-                </a>
-            </div>
-        </div>
-    </div>
-</div>
-
-@section Scripts {
-    <script language="javascript" src="@ViewBag.ControllerJs"></script>
-    <script language="javascript">
-        var vm = new nexplorer.HomeViewModel({
-            latestBlock: @Json.Serialize(Model.LastBlock),
-            primeDiff: @Model.LastPrimeDifficulty,
-            posDiff: @Model.LastPosDifficulty,
-            hashDiff: @Model.LastHashDifficulty
-        });
-    </script>
-}
