@@ -1,1 +1,2 @@
-<script src="@ViewBag.ValidateJs"></script>
+﻿<script src="~/lib/jquery-validation/dist/jquery.validate.min.js"></script>
+<script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"></script>
