@@ -1,3 +1,3 @@
-@using Nexplorer.Web
+﻿@using Nexplorer.Web
 @using Nexplorer.Web.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
